Fix Supabase schema: use DOUBLE PRECISION for time_offset
Critical fix: time_offset must be floating-point, not INTEGER.

Analysis of abracadabra SQLite schema:
- hash table: (hash int, offset real, song_id text)
- song_info table: (artist text, album text, title text, song_id text)

The 'offset' field stores floating-point time values in seconds
(e.g., 1.23456) from the audio spectrogram anchor points.

Changed:
- time_offset INTEGER → time_offset DOUBLE PRECISION

This ensures exact compatibility with abracadabra's fingerprint
matching algorithm which relies on precise time offsets.

+++ supabase_setup.sql
@@ -1,7 +1,11 @@
 -- Supabase Database Setup for Omi Song Recognition
 -- Run this SQL in your Supabase SQL Editor to create the required tables
+--
+-- This schema matches abracadabra's SQLite structure:
+--   SQLite "hash" table → Supabase "fingerprints" table
+--   SQLite "song_info" table → Supabase "songs" table
 
--- Songs table: stores song metadata
+-- Songs table: stores song metadata (matches song_info table)
 CREATE TABLE IF NOT EXISTS songs (
     id SERIAL PRIMARY KEY,
     song_id TEXT UNIQUE NOT NULL,
@@ -11,11 +15,13 @@ CREATE TABLE IF NOT EXISTS songs (
     created_at TIMESTAMP DEFAULT NOW()
 );
 
--- Fingerprints table: stores audio fingerprint hashes
+-- Fingerprints table: stores audio fingerprint hashes (matches hash table)
+-- IMPORTANT: time_offset must be REAL/DOUBLE PRECISION (not INTEGER)
+-- because it stores floating-point time values in seconds (e.g., 1.23456)
 CREATE TABLE IF NOT EXISTS fingerprints (
     id SERIAL PRIMARY KEY,
     hash BIGINT NOT NULL,
-    time_offset INTEGER NOT NULL,
+    time_offset DOUBLE PRECISION NOT NULL,
     song_id TEXT NOT NULL,
     created_at TIMESTAMP DEFAULT NOW(),
     FOREIGN KEY (song_id) REFERENCES songs(song_id) ON DELETE CASCADE
